I should have split this into more commits

+++ bot-controller.py
@@ -1,13 +1,14 @@
 # better bot controller
 
-import re
-import discord
+import discord,  praw , random, re
 from calculester import Calculester
 from env import Env
-from discord.ext import commands
+from calsreddit import CalsReddit
+from discord.ext import commands, tasks
 
 ENV = Env()
-cal = Calculester(ENV.IDs)
+cal = Calculester(ENV.fIDs)
+calsReddit = CalsReddit(ENV)
 
 bot = commands.Bot(command_prefix='!')
 
@@ -21,25 +22,31 @@ async def street_side(ctx):
 
 @bot.command(name='pick-person', help='picks a random person')
 async def pick_person(ctx):
-    await ctx.send(f'<@{cal.pick_friend()}>')
+    if ctx.guild.name == 'Actual friends' :
+        await ctx.send(f'<@{cal.pick_friend()}>')
+    else:
+        await ctx.send('no friends here')
 
 @bot.command(name='fuck', help=':P')
 async def fuck(ctx):
-    await ctx.send(f'fuck <@{cal.pick_friend()}>')
-    await ctx.message.delete()
+    if ctx.guild.name == 'Actual friends' :
+        await ctx.send(f'fuck <@{cal.pick_friend()}>')
+        await ctx.message.delete()
+    else:
+        await ctx.send('im sorry that is nsfw ;)')
 
 @bot.command(name='poop', help='💩')
 async def poop(ctx):
     await ctx.send('💩')
 
-@bot.command(name='unshit')
+@bot.command(name='unshit', help='we all need help')
 async def unshit(ctx, thing: str):
     if thing == 'pants': 
         await ctx.send('💩↩️👖')
     else:
         await ctx.send('💩↩️')
 
-@bot.command(name='tell', help='person to tell followed by message')
+@bot.command(name='tell', help='person to dm followed by message')
 async def tell(ctx, member: discord.Member, *, content):
     channel = await member.create_dm()
     await channel.send(content)
@@ -47,17 +54,51 @@ async def tell(ctx, member: discord.Member, *, content):
 
 @bot.command(name='summon', help='summons the goon squad')
 async def summon(ctx):
-    mem = bot.get_user(cal.nard[0])
-    print(mem)
-    # channel = await mem.create_dm()
-    # await channel.send('summon')
-    await ctx.send('summon test')
-    await ctx.message.delete()
+    if ctx.guild.name == 'Actual friends' :
+        await ctx.send('Summon the goon squad')
+        ats = ''
+        for f in cal.friends :
+            ats += f'<@{f}> '
+        await ctx.send(ats)
 
+@bot.command(name='simp', help='🍆')
+async def simp(ctx):
+    await ctx.message.add_reaction('🍆')
+
+@bot.command(name='porn')
+async def porn(ctx):
+    await ctx.send(calsReddit.getRandomRedditPost('wireporn'))
+
+@bot.command(name='cat', help='r/blep')
+async def cat(ctx):
+    await ctx.send(calsReddit.getRandomRedditPost('Blep'))
+
+@bot.command(name='dog')
+async def dog(ctx):
+    await ctx.send(calsReddit.getRandomRedditPost('rarepuppers'))
+
+# testing function
 @bot.command(name='printer')
 async def printer(ctx):
-    print(ctx.message)
-    await ctx.send('tomato')
-    await ctx.message.delete()
+    # chan = bot.get_channel(734945466982203526)
+    # chan = bot.get_channel(777634476972572693)
+    # print(chan)
+    # print(ctx.channel.id)
+    cal.check_events()
+    await ctx.message.add_reaction('🍆')
 
+@tasks.loop(hours=24)
+async def called_once_a_day():
+    message_channel = bot.get_channel(ENV.genId)
+    print(f"Got channel {message_channel}")
+    todayEvents = cal.check_events()
+    if todayEvents:
+        await message_channel.send(todayEvents)
+
+@called_once_a_day.before_loop
+async def before():
+    await bot.wait_until_ready()
+    print("Finished waiting for cal to wake up")
+
+called_once_a_day.start()
 bot.run(ENV.TOKEN)
